Add email sending feature with FluentEmail integration
This commit introduces a new email sending feature to the application.
It includes configuration settings for SMTP in `appsettings.json`,
new interfaces `IEmailSender` and `IEmailService`, and the
implementation of the `EmailSender` class for sending emails
using Razor templates. A new DTO class `EmailRequestDto`
is added to encapsulate email data. The `ServicesCollection`
class is updated to register the new email services, and
the project file `SnapSell.Infrastructure.csproj` is modified
to include necessary FluentEmail packages.